feat(builder): introduce option traits - add instructions and required to BaseField
